Refactored putting func in dataingest module
